Foi criado uma tabBar para fazer as transaçoes das pages CoffeePage NonCoffeePage PastryPage
