+++ lib/features/pages/coffee_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/src/foundation/key.dart';
+import 'package:flutter/src/widgets/framework.dart';
+import 'package:shopfee_clean_arch/features/components/list-product-widget/list_view_products_widget.dart';
+
+class CoffeePage extends StatelessWidget {
+  const CoffeePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const ListViewProductsWidget();
+  }
+}
