+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:shopfee_clean_arch/features/pages/home_page.dart';
+import 'package:shopfee_clean_arch/features/pages/home_page/home_page.dart';
 
 void main() {
   runApp(const MyApp());
